Squash "comparison between signed and unsigned integer expressions" warnings.

+++ com32/gpllib/e2fsprogs/lib/ext2fs/link.c
@@ -52,9 +52,9 @@ static int link_proc(struct ext2_dir_entry *dirent,
 	 * if so, absorb it into this one.
 	 */
 	next = (struct ext2_dir_entry *) (buf + offset + curr_rec_len);
-	if ((offset + curr_rec_len < blocksize - 8) &&
+	if ((offset + curr_rec_len < (unsigned int)blocksize - 8) &&
 	    (next->inode == 0) &&
-	    (offset + curr_rec_len + next->rec_len <= blocksize)) {
+	    (offset + curr_rec_len + next->rec_len <= (unsigned int)blocksize)) {
 		curr_rec_len += next->rec_len;
 		ls->err = ext2fs_set_rec_len(ls->fs, curr_rec_len, dirent);
 		if (ls->err)
